Refactor data class into shared library.
Refactor the data and serializers so we can pull them into the new centrality class.
