+++ src/centrality/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_executable(centrality centrality.cpp)
+add_executable(centrality centrality.cpp SafegraphMapper.cpp SafegraphMapper.hpp)
 target_link_libraries(centrality PRIVATE map-tile shared Boost::filesystem Boost::regex Boost::program_options spdlog::spdlog yaml-cpp::yaml-cpp)
 target_compile_options(centrality PRIVATE "$<$<CONFIG:DEBUG>:${DEBUG_FLAGS}>")
 target_compile_options(centrality PRIVATE "$<$<CONFIG:RELEASE>:${RELEASE_FLAGS}>")
