+++ src/components/src/server/TileServer.hpp
@@ -6,7 +6,7 @@
 #define MOBILITY_CPP_TILESERVER_HPP
 
 #include "../TileConfiguration.hpp"
-#include "components/data.hpp"
+#include "shared/data.hpp"
 #include <components/JoinedLocation.hpp>
 #include <components/ShapefileWrapper.hpp>
 #include <hpx/include/components.hpp>
