+++ src/shared/src/serializers.cpp
@@ -2,7 +2,7 @@
 // Created by Nicholas Robison on 6/2/20.
 //
 
-#include "serializers.hpp"
+#include "shared/serializers.hpp"
 
 namespace hpx::serialization {
     void
